http files moved to the http module
