+++ Cargo.toml
@@ -49,6 +49,7 @@ pcap-file = { workspace = true }
 hpack-patched = { workspace = true }
 huginn-net-tls = { path = "huginn-net-tls" }
 huginn-net-db = { path = "huginn-net-db" }
+huginn-net-http = { path = "huginn-net-http" }
 
 [dev-dependencies]
 clap = { version = "4.5.47", features = ["derive"] }
