+++ huginn-net-http/Cargo.toml
@@ -3,7 +3,7 @@ name = "huginn-net-http"
 version = "0.1.0"
 edition = "2021"
 rust-version = "1.80"
-description = "HTTP fingerprinting HTTP (p0f-style) analysis for Huginn Net"
+description = "HTTP fingerprinting (p0f-style) analysis for Huginn Net"
 license = "MIT OR Apache-2.0"
 authors = ["Maximiliano Biandratti"]
 repository = "https://github.com/biandratti/huginn-net"
@@ -12,6 +12,9 @@ repository = "https://github.com/biandratti/huginn-net"
 huginn-net-db = { path = "../huginn-net-db" }
 tracing = { workspace = true }
 thiserror = { workspace = true }
-nom = { workspace = true }
+pnet = { workspace = true }
+ttl_cache = { workspace = true }
+hpack-patched = { workspace = true }
+lazy_static = { workspace = true }
