+++ huginn-net-http/src/http2_process.rs
@@ -1,7 +1,7 @@
 use crate::error::HuginnNetError;
 use crate::http::Header;
 use crate::http_common::HttpProcessor;
-use crate::observable_signals::{ObservableHttpRequest, ObservableHttpResponse};
+use crate::observable::{ObservableHttpRequest, ObservableHttpResponse};
 use crate::{http, http2_parser, http_common, http_languages};
 use tracing::debug;
