+++ huginn-net-http/src/lib.rs
@@ -1,7 +1,16 @@
 #![forbid(unsafe_code)]
 
 pub use huginn_net_db as db;
+pub use huginn_net_db::http;
 
-// Placeholder module for future HTTP-specific processing built on top of huginn-net-db
+pub mod http1_parser;
+pub mod http1_process;
+pub mod http2_parser;
+pub mod http2_process;
+pub mod http_common;
+pub mod http_languages;
+pub mod http_process;
 
-
+pub mod display;
+pub mod error;
+pub mod observable;
