+++ src/process.rs
@@ -1,12 +1,12 @@
 use crate::error::HuginnNetError;
-use crate::http_process::{FlowKey, HttpProcessors, ObservableHttpPackage, TcpFlow};
 use crate::observable_signals::ObservableMtu;
 use crate::observable_signals::ObservableTcp;
 use crate::observable_signals::ObservableUptime;
-use crate::observable_signals::{ObservableHttpRequest, ObservableHttpResponse};
 use crate::tcp_process::ObservableTCPPackage;
 use crate::uptime::{Connection, SynData};
-use crate::{http_process, tcp_process, AnalysisConfig};
+use crate::{tcp_process, AnalysisConfig};
+use huginn_net_http::http_process::{FlowKey, HttpProcessors, ObservableHttpPackage, TcpFlow};
+use huginn_net_http::observable::{ObservableHttpRequest, ObservableHttpResponse};
 use huginn_net_tls::{ObservableTlsClient, ObservableTlsPackage};
 use pnet::packet::ip::IpNextHeaderProtocols;
 use pnet::packet::{
@@ -170,7 +170,15 @@ impl IpPacketProcessor for Ipv4Packet<'_> {
         http_processors: &HttpProcessors,
     ) -> Result<ObservableHttpPackage, HuginnNetError> {
         if let Some(packet) = Ipv4Packet::new(data) {
-            http_process::process_http_ipv4(&packet, http_flows, http_processors)
+            huginn_net_http::http_process::process_http_ipv4(&packet, http_flows, http_processors)
+                .map_err(|e| match e {
+                    huginn_net_http::error::HuginnNetError::Parse(msg) => {
+                        HuginnNetError::Parse(msg)
+                    }
+                    huginn_net_http::error::HuginnNetError::UnsupportedProtocol(msg) => {
+                        HuginnNetError::UnsupportedProtocol(msg)
+                    }
+                })
         } else {
             Err(HuginnNetError::UnexpectedPackage(
                 "Invalid IPv4 packet data".to_string(),
@@ -193,8 +201,15 @@ impl IpPacketProcessor for Ipv4Packet<'_> {
 
     fn process_tls_with_data(data: &[u8]) -> Result<ObservableTlsPackage, HuginnNetError> {
         if let Some(packet) = Ipv4Packet::new(data) {
-            huginn_net_tls::process_tls_ipv4(&packet)
-                .map_err(|e| HuginnNetError::Parse(e.to_string()))
+            huginn_net_tls::process_tls_ipv4(&packet).map_err(|e| match e {
+                huginn_net_tls::error::TlsError::Parse(msg) => HuginnNetError::Parse(msg),
+                huginn_net_tls::error::TlsError::UnsupportedProtocol(msg) => {
+                    HuginnNetError::UnsupportedProtocol(msg)
+                }
+                huginn_net_tls::error::TlsError::Unknown => {
+                    HuginnNetError::Parse("Unknown TLS error".to_string())
+                }
+            })
         } else {
             Err(HuginnNetError::UnexpectedPackage(
                 "Invalid IPv4 packet data".to_string(),
@@ -228,7 +243,15 @@ impl IpPacketProcessor for Ipv6Packet<'_> {
         http_processors: &HttpProcessors,
     ) -> Result<ObservableHttpPackage, HuginnNetError> {
         if let Some(packet) = Ipv6Packet::new(data) {
-            http_process::process_http_ipv6(&packet, http_flows, http_processors)
+            huginn_net_http::http_process::process_http_ipv6(&packet, http_flows, http_processors)
+                .map_err(|e| match e {
+                    huginn_net_http::error::HuginnNetError::Parse(msg) => {
+                        HuginnNetError::Parse(msg)
+                    }
+                    huginn_net_http::error::HuginnNetError::UnsupportedProtocol(msg) => {
+                        HuginnNetError::UnsupportedProtocol(msg)
+                    }
+                })
         } else {
             Err(HuginnNetError::UnexpectedPackage(
                 "Invalid IPv6 packet data".to_string(),
@@ -251,8 +274,15 @@ impl IpPacketProcessor for Ipv6Packet<'_> {
 
     fn process_tls_with_data(data: &[u8]) -> Result<ObservableTlsPackage, HuginnNetError> {
         if let Some(packet) = Ipv6Packet::new(data) {
-            huginn_net_tls::process_tls_ipv6(&packet)
-                .map_err(|e| HuginnNetError::Parse(e.to_string()))
+            huginn_net_tls::process_tls_ipv6(&packet).map_err(|e| match e {
+                huginn_net_tls::error::TlsError::Parse(msg) => HuginnNetError::Parse(msg),
+                huginn_net_tls::error::TlsError::UnsupportedProtocol(msg) => {
+                    HuginnNetError::UnsupportedProtocol(msg)
+                }
+                huginn_net_tls::error::TlsError::Unknown => {
+                    HuginnNetError::Parse("Unknown TLS error".to_string())
+                }
+            })
         } else {
             Err(HuginnNetError::UnexpectedPackage(
                 "Invalid IPv6 packet data".to_string(),
